refactor(groups): extract MemberTurnExecutor from GroupConversationService
R1 step 4 of planning/group-collaboration-improvements-plan.md -- the biggest
extraction yet, and the one GroupConversationServiceConcurrencyTest exists
specifically to pin. Moves both executeAgentTurn overloads,
tryResolveMemberToolPause, handleMemberPause, executeGroupMemberTurn,
handleAgentFailure and errorEntry into a new MemberTurnExecutor.

Three design decisions resolved during scoping:
- executeGroupMemberTurn recurses into the facade's own public discuss()/
  cancelDiscussion() for nested GROUP-type members. Resolved by passing
  `this` into MemberTurnExecutor's constructor (safe: it only stores the
  reference, never invokes it during construction).
- propagateDynamicAgentTracking stays on GroupConversationService (widened to
  public static) rather than moving -- DynamicAgentTrackingPropagationTest
  calls it directly by class name, not via reflection, and the plan already
  assigns it to a later step (GroupLifecycleOps).
- Attachment granting reaches back through the same self-reference to the
  facade's grantAndInjectAttachments (widened to public) rather than giving
  MemberTurnExecutor its own IAttachmentStore and duplicating the facade's
  per-call GroupAttachmentBinder construction.

Also widened MemberTurnCancellation/MemberTurnCancelledException to public
(MemberTurnExecutor lives in a different package and references them in its
own signatures).

The reflection sweep needed a second pass -- the first only grepped
getDeclaredMethod("...") and missed this codebase's actual method("...")
test-helper-wrapper pattern, which turned up real dependencies in two files
the first pass said were clean. All six moved methods kept as thin
delegators as a result.

Full 12-class group suite + DynamicAgentTrackingPropagationTest (22 tests,
unmodified) + 4 focused collaborator test classes green on the first run
after compile succeeded, including GroupConversationServiceConcurrencyTest
(8/8). Added a focused MemberTurnExecutorTest for the pure-function methods.

GroupConversationService: 3,977 -> 3,605 lines.
